voice-control/pyproject.toml
[project]
name = "voice-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit>=1.0.23",
    "livekit-agents[openai,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
# Linux-only hardware stack (e.g., Raspberry Pi). Kept optional so macOS dev doesn't fail on evdev.
robot = [
    "trilobot; platform_system == 'Linux'",
]
